requirements.txt
# Core Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.4
SQLAlchemy==2.0.20

# Database
PyMySQL==1.1.0
python-dotenv==1.0.0

# Authentication & Security
PyJWT==2.8.0
Werkzeug==2.3.7

# Validation & Serialization
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0

# Development
python-dateutil==2.8.2
